Change LICENSE to MIT
The reason I originally selected the Unlicense when this project
began was to maximize use of the library. However, since then
I've come to learn that such a license actually prevents use and
contribution by some organizations because Unlicense is not an OSI
approved license and is a little risky legally.

See https://github.com/stacktracejs/stacktrace.js/issues/18